backend/requirements.txt
asgiref==3.11.1
Cython==3.2.4
dj-database-url==3.1.2
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
dotenv==0.9.9
gunicorn==25.3.0
packaging==26.2
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2026.1
whitenoise==6.12.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.11.1
async-generator==1.10
attrs==25.4.0
blinker==1.9.0
build==1.3.0
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==46.0.5
Cython==3.1.4
distro==1.9.0
dj-rest-auth==7.2.0
Django==6.0.2
django-allauth==65.15.0
django-cors-headers==4.9.0
django-environ==0.13.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fastapi==0.116.1
Flask==3.1.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.3
MetaTrader5==5.0.5120
mypy_extensions==1.1.0
mysql-connector-python==9.4.0
mysqlclient==2.2.8
numpy==2.3.1
openai==1.97.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
pillow==12.2.0
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.12.1
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
selenium==4.40.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.5
starlette==0.47.2
ta==0.11.0
TA-Lib==0.6.7
tqdm==4.67.1
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
uvicorn==0.35.0
websocket-client==1.9.0
Werkzeug==3.1.6
wsproto==1.3.2
zipp==3.23.0
